requirements.txt
# Core Dependencies - Updated for Python 3.13 compatibility
pandas>=2.2.0
numpy>=1.26.0
streamlit>=1.29.0

# ML & Embeddings
scikit-learn>=1.3.2
sentence-transformers>=2.2.2
faiss-cpu>=1.9.0
huggingface-hub>=0.16.4
tokenizers>=0.15.0

# LLM Integration - Free/Open Source Options
# Option 1: GPT4All (Completely offline - RECOMMENDED for Mac)
gpt4all>=2.0.2

# Option 2: HuggingFace Transformers (Fully local, no external service)
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# Option 3: Ollama (Run LLMs locally - RECOMMENDED)
# Install Ollama from: https://ollama.ai
# Then run: ollama pull llama2
# Then run: ollama pull mistral
langchain>=0.1.0
langchain-community>=0.0.10

# Visualization
plotly>=5.18.0
jupyter>=1.0.0
ipykernel>=6.27.1
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv>=1.0.0
regex>=2023.10.3
requests>=2.31.0
tqdm==4.65.0

# Optional: For better caching
diskcache>=5.6.3

